# Grouping performance improvements