--- a/src/PopupApp.tsx
+++ b/src/PopupApp.tsx
@@ -1,6 +1,6 @@
 import "./globals.css";
 import Button from "./components/Button";
-import { executeGrouping, unGroupAllTabs } from "./helpers/extensionHelpers";
+import { executeGrouping, ungroupAllTabs } from "./helpers/extensionHelpers";
 import RulesForm from "./components/RulesForm";
 import OptionsForm from "./components/OptionsForm";
 import useRulesRepository from "./hooks/useRulesRepository";
@@ -13,7 +13,7 @@ const PopupApp = () => {
     <div className="popup-app">
       <main className="main-actions">
         <Button onClick={executeGrouping}>Group</Button>
-        <Button onClick={unGroupAllTabs}>Un-Group</Button>
+        <Button onClick={ungroupAllTabs}>Un-Group</Button>
       </main>
       <section>
         <OptionsForm repo={optionsRepo} />