## Perf Notes
Testing in Brave Browser, 2 rules, auto-group, alphabetize
`106` tabs
`105` in proper group before execution

### Baseline
Sort: 150ms
Total: 551ms

### Patch
Sort: 131ms
Total: 142ms